+++ CHANGELOG.md
@@ -29,7 +29,17 @@ calls to `AwsCrypto.encrypt(EncryptRequest)` and `AwsCrypto.decrypt(DecryptReque
    [#165](https://github.com/aws/aws-encryption-sdk-java/pull/165),
    [#168](https://github.com/aws/aws-encryption-sdk-java/pull/168),
    and [#170](https://github.com/aws/aws-encryption-sdk-java/pull/170).
-  
+
+## 1.6.2 -- 2020-05-26
+
+### Patches
+* Validate final frame length does not exceed the frame size in the message header [PR #166](https://github.com/aws/aws-encryption-sdk-java/pull/166)
+* Validate entire ciphertext has been processed before returning [PR #191](https://github.com/aws/aws-encryption-sdk-java/pull/191)
+
+### Maintenance
+* Update AWS Java SDK version from 1.11.561 to 1.11.704. [PR #186](https://github.com/aws/aws-encryption-sdk-java/pull/186)
+* Upgrade Bouncy Castle from 1.61 to 1.65 [PR #179](https://github.com/aws/aws-encryption-sdk-java/pull/179)
+
 ## 1.6.1 -- 2019-10-29
 
 ### Deprecation Warnings